requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
feedgen==1.0.0
PyYAML==6.0.3
python-dateutil==2.9.0.post0
schedule==1.2.2
